Module swap_optimizer::v3::swap

Structs

CurrentState

Current state of a Uniswap V3 swap See SwapState

Step

Steps in a Uniswap V3 swap SeeStepComputation

Tick

Tick data for a Uniswap V3 pool See Tick

Constants

MAX_SQRT_RATIO

MIN_SQRT_RATIO

Functions

get_pool_data

Calls the batch request to get pool data for a Uniswap V3 pool

get_tokens_in_from_tokens_out

Gets the amount of tokens out from the amount of tokens in Note: token0_in and token1_in cannot be both provided See get_tokens_in_from_tokens_out()

get_tokens_out_from_tokens_in

Gets the amount of tokens out from the amount of tokens in Note: token0_in and token1_in cannot be both provided See get_tokens_in_from_tokens_out()

swap

Simulates a Uniswap V3 swap See swap()